- utils添加格式化秒数显示的方法 - 调整Timer代码 - 处理depth loss中缺失深度信息的点(值为inf), 暂行方法见注释

+++ structures/params_3d.py
@@ -14,7 +14,7 @@ class ParamsList():
 
     def add_field(self, field, field_data):
         if not isinstance(field_data, (Calibration, torch.Tensor)):
-            field_data = torch.as_tensor(field_data)
+            field_data = torch.as_tensor(field_data, dtype=torch.float32)
 
         self.extra_fields[field] = field_data
